@@ -13,6 +13,7 @@
   },
   "dependencies": {
     "axios": "^0.15.3",
+    "espree": "^3.5.0",
     "idb": "^2.0.0",
     "lodash": "^4.17.4",
     "lottie-react-native": "^1.2.0",
@@ -21,9 +22,9 @@
     "qrcode.react": "^0.6.1",
     "query-string": "^4.3.4",
     "re-quests": "^1.1.2",
-    "react": "^15.5.4",
+    "react": "^15.6.1",
     "react-css-themr": "^2.1.2",
-    "react-dom": "15.4.2",
+    "react-dom": "^15.6.1",
     "react-native": "^0.42.3",
     "react-native-blur": "^3.0.0-alpha",
     "react-native-camera": "^0.6.0",
@@ -53,63 +54,68 @@
   },
   "devDependencies": {
     "@kadira/storybook": "^2.21.0",
-    "autoprefixer": "6.5.1",
-    "babel-core": "6.17.0",
-    "babel-eslint": "7.1.1",
-    "babel-jest": "^17.0.2",
-    "babel-loader": "6.2.7",
+    "autoprefixer": "7.1.2",
+    "babel-core": "6.25.0",
+    "babel-eslint": "7.2.3",
+    "babel-jest": "20.0.3",
+    "babel-loader": "7.1.1",
     "babel-plugin-flow-react-proptypes": "^3.2.0",
     "babel-plugin-transform-decorators": "^6.24.1",
     "babel-plugin-transform-decorators-legacy": "^1.3.4",
     "babel-preset-es2015": "^6.24.1",
     "babel-preset-react": "^6.24.1",
-    "babel-preset-react-app": "^2.0.1",
+    "babel-preset-react-app": "^3.0.2",
     "babel-preset-react-hmre": "^1.1.1",
     "babel-preset-react-native": "^1.9.1",
-    "case-sensitive-paths-webpack-plugin": "1.1.4",
+    "babel-runtime": "6.23.0",
+    "case-sensitive-paths-webpack-plugin": "2.1.1",
     "chalk": "1.1.3",
     "connect-history-api-fallback": "1.3.0",
     "cross-spawn": "4.0.2",
-    "css-loader": "0.26.0",
+    "css-loader": "0.28.4",
     "detect-port": "1.0.1",
-    "dotenv": "2.0.0",
+    "dotenv": "4.0.0",
     "electron": "^1.6.11",
     "enzyme": "^2.7.1",
-    "eslint": "3.8.1",
-    "eslint-config-react-app": "^0.5.0",
-    "eslint-loader": "1.6.0",
-    "eslint-plugin-flowtype": "2.21.0",
-    "eslint-plugin-import": "2.0.1",
-    "eslint-plugin-jsx-a11y": "2.2.3",
-    "eslint-plugin-react": "6.4.1",
-    "extract-text-webpack-plugin": "1.0.1",
-    "file-loader": "0.9.0",
+    "eslint": "4.4.1",
+    "estraverse": "^4.2.0",
+    "eslint-config-react-app": "^2.0.0",
+    "eslint-loader": "1.9.0",
+    "eslint-plugin-flowtype": "2.35.0",
+    "eslint-plugin-import": "2.7.0",
+    "eslint-plugin-jsx-a11y": "5.1.1",
+    "eslint-plugin-react": "7.1.0",
+    "extract-text-webpack-plugin": "3.0.0",
+    "file-loader": "0.11.2",
     "filesize": "3.3.0",
-    "fs-extra": "0.30.0",
+    "fs-extra": "3.0.1",
     "gzip-size": "3.0.0",
-    "html-webpack-plugin": "2.24.0",
+    "html-webpack-plugin": "2.29.0",
     "http-proxy-middleware": "0.17.2",
     "javascript-natural-sort": "^0.7.1",
-    "jest": "^17.0.2",
+    "jest": "20.0.4",
     "json-loader": "0.5.4",
-    "object-assign": "4.1.0",
+    "object-assign": "4.1.1",
     "path-exists": "2.1.0",
-    "postcss-loader": "1.0.0",
-    "promise": "7.1.1",
+    "postcss-flexbugs-fixes": "3.2.0",
+    "postcss-loader": "2.0.6",
+    "promise": "8.0.1",
     "react-addons-test-utils": "^15.6.0",
-    "react-dev-utils": "^0.4.2",
+    "react-dev-utils": "^3.1.1",
+    "react-error-overlay": "^1.0.10",
     "react-native-dotenv": "^0.1.0",
     "react-test-renderer": "^15.4.2",
     "recursive-readdir": "2.1.0",
     "redux-devtools": "^3.3.2",
     "redux-mock-store": "^1.2.3",
     "strip-ansi": "3.0.1",
-    "style-loader": "0.13.1",
-    "url-loader": "0.5.7",
-    "webpack": "1.14.0",
-    "webpack-dev-server": "1.16.2",
-    "webpack-manifest-plugin": "1.1.0",
-    "whatwg-fetch": "1.0.0"
+    "style-loader": "0.18.2",
+    "sw-precache-webpack-plugin": "0.11.4",
+    "url-loader": "0.5.9",
+    "webpack": "3.5.1",
+    "webpack-dev-server": "2.7.1",
+    "webpack-manifest-plugin": "1.2.1",
+    "whatwg-fetch": "2.0.3"
   },
   "description": "This is the cross platform NoApp.",
   "preset": "react-native",
