@@ -1,4 +1,5 @@
 import React from 'react';
+import PropTypes from 'prop-types';
 
 
 /**
@@ -23,9 +24,9 @@ export function withAutoCompleteProps(WrappedComponent) {
         // contextTypes is a required static property to declare
         // what you want from the context
         static contextTypes = {
-            widget: React.PropTypes.object,
-            setVariable: React.PropTypes.func,
-            enqueue: React.PropTypes.func,
+            widget: PropTypes.object,
+            setVariable: PropTypes.func,
+            enqueue: PropTypes.func,
         };
 
         render() {
