@@ -31,7 +31,7 @@ export function withIconProps(WrappedComponent) {
             }
         };
         static contextTypes = {
-            enqueue: React.PropTypes.func,
+            enqueue: PropTypes.func,
         };
 
         render() {
