@@ -1,4 +1,5 @@
 import React from 'react';
+import PropTypes from 'prop-types';
 import { themr } from 'react-css-themr';
 
 
@@ -23,7 +24,7 @@ export function withTextProps(WrappedComponent) {
     @themr('text')
     class WithTextProps extends React.Component {
         static contextTypes = {
-            enqueue: React.PropTypes.func,
+            enqueue: PropTypes.func,
         };
 
         render() {
