@@ -1,4 +1,5 @@
 import React, { Component } from 'react';
+import PropTypes from 'prop-types';
 import { Modal, Text, TouchableOpacity, View } from 'react-native';
 import { Icon as IconButton } from '../../../re-render';
 
@@ -44,6 +45,6 @@ export default class M extends Component {
 }
 
 M.propTypes = {
-    buttonText: React.PropTypes.string.isRequired,
-    contentLabel: React.PropTypes.string.isRequired,
+    buttonText: PropTypes.string.isRequired,
+    contentLabel: PropTypes.string.isRequired,
 };
