@@ -1,47 +1,50 @@
-var autoprefixer = require('autoprefixer');
-var webpack = require('webpack');
-var HtmlWebpackPlugin = require('html-webpack-plugin');
-var ExtractTextPlugin = require('extract-text-webpack-plugin');
-var ManifestPlugin = require('webpack-manifest-plugin');
-var InterpolateHtmlPlugin = require('react-dev-utils/InterpolateHtmlPlugin');
-var url = require('url');
-var paths = require('./paths');
-var getClientEnvironment = require('./env');
+'use strict';
 
+const autoprefixer = require('autoprefixer');
+const path = require('path');
+const webpack = require('webpack');
+const HtmlWebpackPlugin = require('html-webpack-plugin');
+const ExtractTextPlugin = require('extract-text-webpack-plugin');
+const ManifestPlugin = require('webpack-manifest-plugin');
+const InterpolateHtmlPlugin = require('react-dev-utils/InterpolateHtmlPlugin');
+const SWPrecacheWebpackPlugin = require('sw-precache-webpack-plugin');
+const eslintFormatter = require('react-dev-utils/eslintFormatter');
+const ModuleScopePlugin = require('react-dev-utils/ModuleScopePlugin');
+const paths = require('./paths');
+const getClientEnvironment = require('./env');
 
-function ensureSlash(path, needsSlash) {
-    var hasSlash = path.endsWith('/');
-    if (hasSlash && !needsSlash) {
-        return path.substr(path, path.length - 1);
-    } else if (!hasSlash && needsSlash) {
-        return path + '/';
-    } else {
-        return path;
-    }
-}
-
-// We use "homepage" field to infer "public path" at which the app is served.
-// Webpack needs to know it to put the right <script> hrefs into HTML even in
-// single-page apps that may serve index.html for nested URLs like /todos/42.
-// We can't use a relative path in HTML because we don't want to load something
-// like /todos/42/static/js/bundle.7289d.js. We have to know the root.
-var homepagePath = require(paths.appPackageJson).homepage;
-var homepagePathname = homepagePath ? url.parse(homepagePath).pathname : '/';
 // Webpack uses `publicPath` to determine where the app is being served from.
 // It requires a trailing slash, or the file assets will get an incorrect path.
-var publicPath = ensureSlash(homepagePathname, true);
+const publicPath = paths.servedPath;
+// Some apps do not use client-side routing with pushState.
+// For these, "homepage" can be set to "." to enable relative asset paths.
+const shouldUseRelativeAssetPaths = publicPath === './';
+// Source maps are resource heavy and can cause out of memory issue for large source files.
+const shouldUseSourceMap = process.env.GENERATE_SOURCEMAP !== 'false';
 // `publicUrl` is just like `publicPath`, but we will provide it to our app
 // as %PUBLIC_URL% in `index.html` and `process.env.PUBLIC_URL` in JavaScript.
-// Omit trailing slash as %PUBLIC_PATH%/xyz looks better than %PUBLIC_PATH%xyz.
-var publicUrl = ensureSlash(homepagePathname, false);
+// Omit trailing slash as %PUBLIC_URL%/xyz looks better than %PUBLIC_URL%xyz.
+const publicUrl = publicPath.slice(0, -1);
 // Get environment variables to inject into our app.
-var env = getClientEnvironment(publicUrl);
+const env = getClientEnvironment(publicUrl);
 
 // Assert this just to be safe.
 // Development builds of React are slow and not intended for production.
-if (env['process.env'].NODE_ENV !== '"production"') {
+if (env.stringified['process.env'].NODE_ENV !== '"production"') {
     throw new Error('Production builds must have NODE_ENV=production.');
 }
+
+// Note: defined here because it will be used more than once.
+const cssFilename = 'static/css/[name].[contenthash:8].css';
+
+// ExtractTextPlugin expects the build output to be flat.
+// (See https://github.com/webpack-contrib/extract-text-webpack-plugin/issues/27)
+// However, our output is structured with css, js and media folders.
+// To have this structure working with relative paths, we have to use custom options.
+const extractTextPluginOptions = shouldUseRelativeAssetPaths
+    ? // Making sure that the publicPath goes back to to build folder.
+    {publicPath: Array(cssFilename.split('/').length).join('../')}
+    : {};
 
 // This is the production configuration.
 // It compiles slowly and is focused on producing a fast and minimal bundle.
@@ -51,12 +54,9 @@ module.exports = {
     bail: true,
     // We generate sourcemaps in production. This is slow but gives good results.
     // You can exclude the *.map files from the build during deployment.
-    devtool: 'source-map',
+    devtool: shouldUseSourceMap ? 'source-map' : false,
     // In production, we only want to load the polyfills and the app code.
-    entry: [
-        require.resolve('./polyfills'),
-        paths.appIndexJs
-    ],
+    entry: [require.resolve('./polyfills'), paths.appIndexJs],
     output: {
         // The build folder.
         path: paths.appBuild,
@@ -66,134 +66,179 @@ module.exports = {
         filename: 'static/js/[name].[chunkhash:8].js',
         chunkFilename: 'static/js/[name].[chunkhash:8].chunk.js',
         // We inferred the "public path" (such as / or /my-project) from homepage.
-        publicPath: publicPath
+        publicPath: publicPath,
+        // Point sourcemap entries to original disk location (format as URL on Windows)
+        devtoolModuleFilenameTemplate: info =>
+            path
+                .relative(paths.appSrc, info.absoluteResourcePath)
+                .replace(/\\/g, '/'),
     },
     resolve: {
         // This allows you to set a fallback for where Webpack should look for modules.
-        // We read `NODE_PATH` environment variable in `paths.js` and pass paths here.
-        // We use `fallback` instead of `root` because we want `node_modules` to "win"
-        // if there any conflicts. This matches Node resolution mechanism.
+        // We placed these paths second because we want `node_modules` to "win"
+        // if there are any conflicts. This matches Node resolution mechanism.
         // https://github.com/facebookincubator/create-react-app/issues/253
-        fallback: paths.nodePaths,
+        modules: ['node_modules', paths.appNodeModules].concat(
+            // It is guaranteed to exist because we tweak it in `env.js`
+            process.env.NODE_PATH.split(path.delimiter).filter(Boolean)
+        ),
         // These are the reasonable defaults supported by the Node ecosystem.
         // We also include JSX as a common component filename extension to support
         // some tools, although we do not recommend using it, see:
         // https://github.com/facebookincubator/create-react-app/issues/290
-        extensions: ['.js', '.json', '.jsx', ''],
+        // `web` extension prefixes have been added for better support
+        // for React Native Web.
+        extensions: ['.web.js', '.js', '.json', '.web.jsx', '.jsx'],
         alias: {
+
             // Support React Native Web
             // https://www.smashingmagazine.com/2016/08/a-glimpse-into-the-future-with-react-native-for-web/
             'react-native': 'react-native-web',
-            'react-router-native': 'react-router'
-        }
-    },
-
-    module: {
-        // First, run the linter.
-        // It's important to do this before Babel processes the JS.
-        preLoaders: [
-            {
-                test: /\.(js|jsx)$/,
-                loader: 'eslint',
-                include: [paths.appIndexJs, paths.appSrc]
-            }
+        },
+        plugins: [
+            // Prevents users from importing files from outside of src/ (or node_modules/).
+            // This often causes confusion because we only process files within src/ with babel.
+            // To fix this, we prevent you from importing files out of src/ -- if you'd like to,
+            // please link the files into your node_modules/ and let module-resolution kick in.
+            // Make sure your source files are compiled, as they will not be processed in any way.
+            new ModuleScopePlugin(paths.appSrc, [paths.appPackageJson]),
         ],
-        loaders: [
-            // Default loader: load all assets that are not handled
-            // by other loaders with the url loader.
-            // Note: This list needs to be updated with every change of extensions
-            // the other loaders match.
-            // E.g., when adding a loader for a new supported file extension,
-            // we need to add the supported extension to this loader too.
-            // Add one new line in `exclude` for each loader.
-            //
-            // "file" loader makes sure those assets end up in the `build` folder.
-            // When you `import` an asset, you get its filename.
-            // "url" loader works just like "file" loader but it also embeds
-            // assets smaller than specified size as data URLs to avoid requests.
-            {
-                exclude: [
-                    /\.html$/,
-                    /\.(js|jsx)$/,
-                    /\.css$/,
-                    /\.json$/,
-                    /\.svg$/
-                ],
-                loader: 'url',
-                query: {
-                    limit: 10000,
-                    name: 'static/media/[name].[hash:8].[ext]'
-                }
-            },
-            // Process JS with Babel.
+    },
+    module: {
+        strictExportPresence: true,
+        rules: [
+            // TODO: Disable require.ensure as it's not a standard language feature.
+            // We are waiting for https://github.com/facebookincubator/create-react-app/issues/2176.
+            // { parser: { requireEnsure: false } },
+
+            // First, run the linter.
+            // It's important to do this before Babel processes the JS.
             {
                 test: /\.(js|jsx)$/,
-                include: [paths.appIndexJs, paths.appSrc],
-                exclude: /node_modules\/(?!(react-router-native)\/).*/,
-                loader: 'babel',
+                enforce: 'pre',
+                use: [
+                    {
+                        options: {
+                            formatter: eslintFormatter,
+                            eslintPath: require.resolve('eslint'),
 
+                        },
+                        loader: require.resolve('eslint-loader'),
+                    },
+                ],
+                include: paths.appSrc,
             },
             {
-                test: /\.js$/,
-                exclude: /node_modules\/(?!(react-native-elements|react-native-tab-view|react-native-vector-icons|react-native-drawer|react-native-scrollable-tab-view|react-native-popup-dialog|react-native-side-menu|react-native-tab-navigator|react-native-grid-component)\/).*/,
-                loader: 'babel-loader'
+                // "oneOf" will traverse all following loaders until one will
+                // match the requirements. When no loader matches it will fall
+                // back to the "file" loader at the end of the loader list.
+                oneOf: [
+                    // "url" loader works just like "file" loader but it also embeds
+                    // assets smaller than specified size as data URLs to avoid requests.
+                    {
+                        test: [/\.bmp$/, /\.gif$/, /\.jpe?g$/, /\.png$/],
+                        loader: require.resolve('url-loader'),
+                        options: {
+                            limit: 10000,
+                            name: 'static/media/[name].[hash:8].[ext]',
+                        },
+                    },
+                    // Process JS with Babel.
+                    {
+                        test: /\.(js|jsx)$/,
+                        include: [
+                            paths.appIndexJs,
+                            paths.appSrc,
+                            /node_modules\/(react-native-elements|react-native-tab-view|react-native-vector-icons|react-native-drawer|react-native-scrollable-tab-view|react-native-popup-dialog|react-native-side-menu|react-native-tab-navigator|react-native-grid-component)\/.*/
+                        ],
+                        loader: require.resolve('babel-loader'),
+                        options: {
+                            compact: true,
+                        },
+                    },
+                    // The notation here is somewhat confusing.
+                    // "postcss" loader applies autoprefixer to our CSS.
+                    // "css" loader resolves paths in CSS and adds assets as dependencies.
+                    // "style" loader normally turns CSS into JS modules injecting <style>,
+                    // but unlike in development configuration, we do something different.
+                    // `ExtractTextPlugin` first applies the "postcss" and "css" loaders
+                    // (second argument), then grabs the result CSS and puts it into a
+                    // separate file in our build process. This way we actually ship
+                    // a single CSS file in production instead of JS code injecting <style>
+                    // tags. If you use code splitting, however, any async bundles will still
+                    // use the "style" loader inside the async code so CSS from them won't be
+                    // in the main CSS file.
+                    {
+                        test: /\.css$/,
+                        loader: ExtractTextPlugin.extract(
+                            Object.assign(
+                                {
+                                    fallback: require.resolve('style-loader'),
+                                    use: [
+                                        {
+                                            loader: require.resolve('css-loader'),
+                                            options: {
+                                                importLoaders: 1,
+                                                minimize: true,
+                                                sourceMap: shouldUseSourceMap,
+                                            },
+                                        },
+                                        {
+                                            loader: require.resolve('postcss-loader'),
+                                            options: {
+                                                // Necessary for external CSS imports to work
+                                                // https://github.com/facebookincubator/create-react-app/issues/2677
+                                                ident: 'postcss',
+                                                plugins: () => [
+                                                    require('postcss-flexbugs-fixes'),
+                                                    autoprefixer({
+                                                        browsers: [
+                                                            '>1%',
+                                                            'last 4 versions',
+                                                            'Firefox ESR',
+                                                            'not ie < 9', // React doesn't support IE8 anyway
+                                                        ],
+                                                        flexbox: 'no-2009',
+                                                    }),
+                                                ],
+                                            },
+                                        },
+                                    ],
+                                },
+                                extractTextPluginOptions
+                            )
+                        ),
+                        // Note: this won't work without `new ExtractTextPlugin()` in `plugins`.
+                    },
+                    // "file" loader makes sure assets end up in the `build` folder.
+                    // When you `import` an asset, you get its filename.
+                    // This loader don't uses a "test" so it will catch all modules
+                    // that fall through the other loaders.
+                    {
+                        loader: require.resolve('file-loader'),
+                        // Exclude `js` files to keep "css" loader working as it injects
+                        // it's runtime that would otherwise processed through "file" loader.
+                        // Also exclude `html` and `json` extensions so they get processed
+                        // by webpacks internal loaders.
+                        exclude: [/\.js$/, /\.html$/, /\.json$/],
+                        options: {
+                            name: 'static/media/[name].[hash:8].[ext]',
+                        },
+                    },
+                    // ** STOP ** Are you adding a new loader?
+                    // Make sure to add the new loader(s) before the "file" loader.
+                ],
             },
-            // The notation here is somewhat confusing.
-            // "postcss" loader applies autoprefixer to our CSS.
-            // "css" loader resolves paths in CSS and adds assets as dependencies.
-            // "style" loader normally turns CSS into JS modules injecting <style>,
-            // but unlike in development configuration, we do something different.
-            // `ExtractTextPlugin` first applies the "postcss" and "css" loaders
-            // (second argument), then grabs the result CSS and puts it into a
-            // separate file in our build process. This way we actually ship
-            // a single CSS file in production instead of JS code injecting <style>
-            // tags. If you use code splitting, however, any async bundles will still
-            // use the "style" loader inside the async code so CSS from them won't be
-            // in the main CSS file.
-            {
-                test: /\.css$/,
-                loader: ExtractTextPlugin.extract('style', 'css?importLoaders=1!postcss')
-                // Note: this won't work without `new ExtractTextPlugin()` in `plugins`.
-            },
-            // JSON is not enabled by default in Webpack but both Node and Browserify
-            // allow it implicitly so we also enable it.
-            {
-                test: /\.json$/,
-                loader: 'json'
-            },
-            // "file" loader for svg
-            {
-                test: /\.svg$/,
-                loader: 'file',
-                query: {
-                    name: 'static/media/[name].[hash:8].[ext]'
-                }
-            }
-        ]
-    },
-
-    // We use PostCSS for autoprefixing only.
-    postcss: function () {
-        return [
-            autoprefixer({
-                browsers: [
-                    '>1%',
-                    'last 4 versions',
-                    'Firefox ESR',
-                    'not ie < 9', // React doesn't support IE8 anyway
-                ]
-            }),
-        ];
+        ],
     },
     plugins: [
-        new webpack.ContextReplacementPlugin(/moment[\/\\]locale$/, /en/),
-        // Makes the public URL available as %PUBLIC_URL% in index.html, e.g.:
+        new webpack.ContextReplacementPlugin(/moment[/\\]locale$/, /en/),
+        // Makes some environment variables available in index.html.
+        // The public URL is available as %PUBLIC_URL% in index.html, e.g.:
         // <link rel="shortcut icon" href="%PUBLIC_URL%/favicon.ico">
         // In production, it will be an empty string unless you specify "homepage"
         // in `package.json`, in which case it will be the pathname of that URL.
-        new InterpolateHtmlPlugin({
-            PUBLIC_URL: publicUrl
-        }),
+        new InterpolateHtmlPlugin(env.raw),
         // Generates an `index.html` file with the <script> injected.
         new HtmlWebpackPlugin({
             inject: true,
@@ -208,46 +253,85 @@ module.exports = {
                 keepClosingSlash: true,
                 minifyJS: true,
                 minifyCSS: true,
-                minifyURLs: true
-            }
+                minifyURLs: true,
+            },
         }),
         // Makes some environment variables available to the JS code, for example:
         // if (process.env.NODE_ENV === 'production') { ... }. See `./env.js`.
         // It is absolutely essential that NODE_ENV was set to production here.
         // Otherwise React will be compiled in the very slow development mode.
-        new webpack.DefinePlugin(env),
-        // This helps ensure the builds are consistent if source hasn't changed:
-        new webpack.optimize.OccurrenceOrderPlugin(),
-        // Try to dedupe duplicated modules, if any:
-        new webpack.optimize.DedupePlugin(),
+        new webpack.DefinePlugin(env.stringified),
         // Minify the code.
         new webpack.optimize.UglifyJsPlugin({
             compress: {
-                screw_ie8: true, // React doesn't support IE8
-                warnings: false
-            },
-            mangle: {
-                screw_ie8: true
+                warnings: false,
+                // Disabled because of an issue with Uglify breaking seemingly valid code:
+                // https://github.com/facebookincubator/create-react-app/issues/2376
+                // Pending further investigation:
+                // https://github.com/mishoo/UglifyJS2/issues/2011
+                comparisons: false,
             },
             output: {
                 comments: false,
-                screw_ie8: true
-            }
+                // Turned on because emoji and regex is not minified properly using default
+                // https://github.com/facebookincubator/create-react-app/issues/2488
+                ascii_only: true,
+            },
+            sourceMap: shouldUseSourceMap,
         }),
         // Note: this won't work without ExtractTextPlugin.extract(..) in `loaders`.
-        new ExtractTextPlugin('static/css/[name].[contenthash:8].css'),
+        new ExtractTextPlugin({
+            filename: cssFilename,
+        }),
         // Generate a manifest file which contains a mapping of all asset filenames
         // to their corresponding output file so that tools can pick it up without
         // having to parse `index.html`.
         new ManifestPlugin({
-            fileName: 'asset-manifest.json'
-        })
+            fileName: 'asset-manifest.json',
+        }),
+        // Generate a service worker script that will precache, and keep up to date,
+        // the HTML & assets that are part of the Webpack build.
+        new SWPrecacheWebpackPlugin({
+            // By default, a cache-busting query parameter is appended to requests
+            // used to populate the caches, to ensure the responses are fresh.
+            // If a URL is already hashed by Webpack, then there is no concern
+            // about it being stale, and the cache-busting can be skipped.
+            dontCacheBustUrlsMatching: /\.\w{8}\./,
+            filename: 'service-worker.js',
+            logger(message) {
+                if (message.indexOf('Total precache size is') === 0) {
+                    // This message occurs for every build and is a bit too noisy.
+                    return;
+                }
+                if (message.indexOf('Skipping static resource') === 0) {
+                    // This message obscures real errors so we ignore it.
+                    // https://github.com/facebookincubator/create-react-app/issues/2612
+                    return;
+                }
+                console.log(message);
+            },
+            minify: true,
+            // For unknown URLs, fallback to the index page
+            navigateFallback: publicUrl + '/index.html',
+            // Ignores URLs starting from /__ (useful for Firebase):
+            // https://github.com/facebookincubator/create-react-app/issues/2237#issuecomment-302693219
+            navigateFallbackWhitelist: [/^(?!\/__).*/],
+            // Don't precache sourcemaps (they're large) and build asset manifest:
+            staticFileGlobsIgnorePatterns: [/\.map$/, /asset-manifest\.json$/],
+        }),
+        // Moment.js is an extremely popular library that bundles large locale files
+        // by default due to how Webpack interprets its code. This is a practical
+        // solution that requires the user to opt into importing specific locales.
+        // https://github.com/jmblog/how-to-optimize-momentjs-with-webpack
+        // You can remove this if you don't use Moment.js:
+        new webpack.IgnorePlugin(/^\.\/locale$/, /moment$/),
     ],
     // Some libraries import Node modules but don't use them in the browser.
     // Tell Webpack to provide empty mocks for them so importing them works.
     node: {
+        dgram: 'empty',
         fs: 'empty',
         net: 'empty',
-        tls: 'empty'
-    }
+        tls: 'empty',
+    },
 };
